Improve alignment of inline charts with text
The svg text components cant be at the bottom of the svg container as
small parts of the text might get cut off. The hieght is therefore
increased by a small amount, so that nothing is placed at the absolute
bottom. This missalignment is then fixed using a transfrom on the svg
parent element